refactor: add missing imports (part 5)
