+++ src/renderer/app/components/ContextMenu/Page/index.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { observer } from 'mobx-react-lite';
 
 import store from '~/renderer/app/store';
-import { icons } from '~/renderer/constants';
 import { MenuContainer, MenuItem } from '..';
 import { MenuDivider } from '../style';
+import { icons } from '~/renderer/constants/icons';
 
 const onRefresh = () => {
   store.pages.current.fetchFiles();
