+++ src/renderer/mixins/images.ts
@@ -1,4 +1,7 @@
-export const centerIcon = (size: number | 'contain' = 'contain', useMask = false) => {
+export const centerIcon = (
+  size: number | 'contain' = 'contain',
+  useMask = false,
+) => {
   let s: string = size.toString();
 
   if (typeof size === 'number') s += 'px';
